mess up collisions
now it only sends the positions when the object collides (that way there is no jittering when moving).
But, this exposed a new problem. The clients end is animated faster than the server end (the longer you travel the further you rubber band through a wall) I am not sure how to fix this. The client receives the positions as soon as they are sent and still rubber bands because the client is just animated faster.